remote_world_manager split-out
