+++ shared/src/connection/base_connection.rs
@@ -3,6 +3,7 @@ use std::{hash::Hash, net::SocketAddr};
 use naia_serde::{BitWriter, Serde};
 use naia_socket_shared::Instant;
 
+use crate::world::remote::remote_world_reader::RemoteWorldReader;
 use crate::{
     backends::Timer,
     messages::{channels::channel_kinds::ChannelKinds, message_manager::MessageManager},
@@ -26,6 +27,7 @@ pub struct BaseConnection<E: Copy + Eq + Hash + Send + Sync> {
     pub message_manager: MessageManager,
     pub host_world_manager: HostWorldManager<E>,
     pub remote_world_manager: RemoteWorldManager<E>,
+    pub remote_world_reader: RemoteWorldReader<E>,
     pub local_world_manager: LocalWorldManager<E>,
     heartbeat_timer: Timer,
     timeout_timer: Timer,
@@ -48,6 +50,7 @@ impl<E: Copy + Eq + Hash + Send + Sync> BaseConnection<E> {
             message_manager: MessageManager::new(host_type, channel_kinds),
             host_world_manager: HostWorldManager::new(address, global_world_manager),
             remote_world_manager: RemoteWorldManager::new(),
+            remote_world_reader: RemoteWorldReader::new(),
             local_world_manager: LocalWorldManager::new(),
         }
     }
