+++ shared/src/world/remote/mod.rs
@@ -2,3 +2,4 @@ pub mod entity_action_event;
 pub mod entity_event;
 pub mod entity_waitlist;
 pub mod remote_world_manager;
+pub mod remote_world_reader;
